requirements.txt
# === Core Python packages ===
numpy>=1.25
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.66

# === Image processing & augmentation ===
opencv-python>=4.10
Pillow>=10.0
albumentations>=1.4
scikit-image>=0.22
imbalanced-learn>=0.12

# === Deep Learning ===
torch>=2.1
torchvision>=0.16
torchaudio>=2.1

# === Gradient-based explainability ===
shap>=0.45

# === Traditional ML models ===
xgboost>=2.0
joblib>=1.3

# === Web app & backend ===
Flask>=3.0
Flask-Cors>=4.0
reportlab>=4.0   # For PDF generation

# === Notebook & development tools ===
jupyterlab>=4.0

# === Optional (for visualization, analytics) ===
plotly>=5.20
